test(Radio): add new tests
